+++ src/pages/_app.js
@@ -1,0 +1,7 @@
+import '../styles/globals.css';
+
+function MyApp({ Component, pageProps }) {
+  return <Component {...pageProps} />;
+}
+
+export default MyApp;
